requirements.txt
Flask
requests
hazelcast-python-client
python-consul